Make the fetch user signUp & logIn in one slice
