+++ src/redux/configureStore.js
@@ -1,16 +1,15 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createLogger } from 'redux-logger';
 import registerReducer from './users/registerSlice';
-import loginReducer from './users/loginSlice';
-import airlineReducer from './data/airlines';
+// import loginReducer from './users/loginSlice';
+import airlineReducer from './airlines/airlines';
 
 const logger = createLogger();
 
 const store = configureStore({
   reducer: {
     airlines: airlineReducer,
-    register: registerReducer,
-    login: loginReducer,
+    userInfo: registerReducer,
   },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 });
