+++ src/redux/users/registerSlice.js
@@ -2,8 +2,8 @@ import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import baseURL from '../api';
 
 const initialState = {
-  loading: false,
   register: [],
+  login: [],
   error: '',
 };
 
@@ -18,26 +18,38 @@ export const fetchNewUser = createAsyncThunk('register/fetchNewUser', (newUser) 
   return res;
 });
 
+export const fetchCurrentUser = createAsyncThunk('login/fetchCurrentUser', (currentUser) => {
+  const res = fetch(`${baseURL}/login`, {
+    method: 'POST',
+    headers: {
+      'Content-type': 'application/json',
+    },
+    body: JSON.stringify(currentUser),
+  }).then((res) => res.json());
+  return res;
+});
+
 const registerSlice = createSlice({
   name: 'register',
   initialState,
   extraReducers: (builder) => {
-    builder.addCase(fetchNewUser.pending, (state) => {
-      const stateLoading = state;
-      stateLoading.loading = true;
-    });
-
     builder.addCase(fetchNewUser.fulfilled, (state, action) => {
       const stateFulfilled = state;
-      stateFulfilled.loading = false;
       stateFulfilled.register = action.payload;
-      stateFulfilled.error = '';
     });
 
     builder.addCase(fetchNewUser.rejected, (state, action) => {
       const stateRejected = state;
-      stateRejected.loading = false;
-      stateRejected.register = [];
+      stateRejected.error = action.error.message;
+    });
+
+    builder.addCase(fetchCurrentUser.fulfilled, (state, action) => {
+      const stateFulfilled = state;
+      stateFulfilled.login = action.payload;
+    });
+
+    builder.addCase(fetchCurrentUser.rejected, (state, action) => {
+      const stateRejected = state;
       stateRejected.error = action.error.message;
     });
   },
